requirements_gpu.txt
imageio==2.3.0
matplotlib==2.2.2
numpy==1.14.2
scikit_image==0.13.1
scipy==1.0.1
tensorflow-gpu==1.7.0
tensorboard==1.7.0
tensorlayer>=1.8.4
scikit-learn==0.19.1